fix: move cards controller to avoid conflicts
