+++ app/views/avo/dashboards/cards/_chartkick_card.html.erb
