+++ app/views/avo/dashboards/cards/_metric_card.html.erb
