+++ app/views/avo/dashboards/cards/chartkick_missing.html.erb
