+++ app/views/avo/dashboards/cards/show.html.erb
